pyproject.toml
[tool.poetry]
name = "redsuggested"
version = "0.1.0"
description = ""
authors = ["Chat Sumlin <chat@chatsumlin.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
tautulli = "^4.3.4.2140"
pylistenbrainz = "^0.5.1"
python-dotenv = "^1.0.1"
gotify = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
